+++ app/Http/Controllers/TaskStatusController.php
@@ -44,9 +44,11 @@ class TaskStatusController extends Controller
     public function store(StoreTaskStatusRequest $request)
     {
         $data = $request->validated();
-        $request->user()->taskStatuses()->create($data);
 
-        session()->flash('message', 'New status created successfully');
+        $taskStatus = new TaskStatus($data);
+        $taskStatus->save();
+
+        session()->flash('success', "New status created successfully");
 
         return redirect()
             ->route('task_statuses.index');
@@ -69,7 +71,7 @@ class TaskStatusController extends Controller
 
         $taskStatus->fill($data)->save();
 
-        session()->flash('message', 'Status edited successfully');
+        session()->flash('success', "Status edited successfully");
 
         return redirect()
             ->route('task_statuses.index');
@@ -80,10 +82,13 @@ class TaskStatusController extends Controller
      */
     public function destroy(TaskStatus $taskStatus)
     {
-        $taskStatusName = $taskStatus->name;
+        if (!empty($taskStatus->tasks()->first())) {
+            session()->flash('error', "Не удалось удалить статус");
+            return back();
+        }
         $taskStatus->delete();
 
-        session()->flash('message', "Status \"{$taskStatusName}\" deleted successfully");
+        session()->flash('success', "Status deleted successfully");
 
         return redirect()
             ->route('task_statuses.index');
